Migrated rest test to new test suites and updated action YAML
