--- .github/actions/rest-test/action.yaml
@@ -1,79 +1,0 @@
----
-name: Run rest test
-description: Validate that rest endpoints return expected results
-runs:
-  using: composite
-  steps:
-    - uses: actions/setup-node@v6
-      with:
-        node-version: 20
-
-    - name: Install dependencies (Windows)
-      if: runner.os == 'Windows'
-      working-directory: tests/rest
-      shell: cmd
-      run: npm install
-
-    - name: Install dependencies (Linux)
-      if: runner.os == 'Linux'
-      working-directory: tests/rest
-      shell: /usr/bin/bash {0}
-      env:
-        PATH: /usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin
-      run: npm install
-
-    - name: Start NSClient++ (Windows)
-      if: runner.os == 'Windows'
-      working-directory: tmp/nscp
-      shell: cmd
-      run: |
-        del nsclient.log
-        cmd /c "start nscp test --settings ..\..\tests\rest\nsclient.ini"
-
-    - name: Start NSClient++ (Linux)
-      if: runner.os == 'Linux'
-      shell: /usr/bin/bash {0}
-      env:
-        PATH: /usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin
-      run: |
-        nscp test --settings tests/rest/nsclient.ini > /tmp/nsclient.log 2>&1 &
-
-    - name: Run tests (Windows)
-      if: always() && runner.os == 'Windows'
-      working-directory: tests/rest
-      shell: sh
-      run: npm run test
-    - name: Run tests (Linux)
-      if: runner.os == 'Linux'
-      working-directory: tests/rest
-      shell: /usr/bin/bash {0}
-      env:
-        PATH: /usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin
-      run: npm run test
-
-    - name: Show logs (Windows)
-      working-directory: tmp/nscp
-      if: always() && runner.os == 'Windows'
-      shell: sh
-      run: cat nsclient.log || true
-
-    - name: Show logs (Linux)
-      if: always() && runner.os == 'Linux'
-      shell: /usr/bin/bash {0}
-      env:
-        PATH: /usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin
-      run: cat /tmp/nsclient.log || true
-
-    - name: Stop NSClient++ (Windows)
-      if: always() && runner.os == 'Windows'
-      working-directory: tmp/nscp
-      shell: cmd
-      run: taskkill /F /im nscp.exe
-
-    - name: Stop NSClient++ (Linux)
-      if: always() && runner.os == 'Linux'
-      shell: /usr/bin/bash {0}
-      env:
-        PATH: /usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin
-      run: pkill -f nscp || true
-
