+++ tests/package.json
@@ -5,15 +5,18 @@
   "description": "Cross-platform scenario tests for NSClient++ (replaces tests/<proto>/run-test.bat).",
   "scripts": {
     "test": "jest --runInBand",
-    "test:one": "jest --runInBand --testPathPattern"
+    "test:one": "jest --runInBand --testPathPattern",
+    "test:no-docker": "NSCP_SKIP_DOCKER=1 jest --runInBand"
   },
   "devDependencies": {
     "@types/jest": "^29.5.14",
     "@types/node": "^20.14.10",
     "@types/node-forge": "^1.3.11",
+    "@types/supertest": "^6.0.3",
     "execa": "^5.1.1",
     "jest": "^29.7.0",
     "node-forge": "^1.3.1",
+    "supertest": "^7.2.2",
     "testcontainers": "^10.13.2",
     "ts-jest": "^29.2.5",
     "typescript": "^5.6.3"
