+++ tests/rest-aliases-v2.test.ts
@@ -1,17 +1,36 @@
-import { describe, expect, it } from "@jest/globals";
-import request = require("supertest");
+/**
+ * REST alias-endpoint scenarios — migrated from tests/rest/aliases-v2.test.ts.
+ *
+ * Verifies the QUERY_ALIAS items registered by CheckHelpers under
+ * [/settings/check helpers/alias] (see setupRestNscp) show up on /aliases
+ * with the expected shape, and that plain queries don't leak into the
+ * alias listing.
+ */
+import request from "supertest";
+import { NscpInstance, REST_URL, setupRestNscp } from "@fixtures/index";
 
-const URL = "https://127.0.0.1:8443";
+jest.setTimeout(900_000);
 
-// Aliases are defined under [/settings/check helpers/alias] in nsclient.ini -
-// the CheckHelpers module registers them as QUERY_ALIAS items at load time.
+// Aliases are defined under [/settings/check helpers/alias] in the shared
+// fixture — the CheckHelpers module registers them as QUERY_ALIAS items at
+// load time.
 const expectedAliases = ["mock_alias", "echo_alias"];
 
-describe("aliases (v2)", () => {
+describe("REST aliases (v2)", () => {
+  let nscp: NscpInstance;
   let key: string | undefined = undefined;
 
+  beforeAll(async () => {
+    nscp = new NscpInstance();
+    await setupRestNscp(nscp);
+  });
+
+  afterAll(async () => {
+    await nscp?.stop();
+  });
+
   it("can login", async () => {
-    await request(URL)
+    await request(REST_URL)
       .get("/api/v2/login")
       .auth("admin", "default-password")
       .trustLocalhost(true)
@@ -24,7 +43,7 @@ describe("aliases (v2)", () => {
   });
 
   it("list aliases returns the registered QUERY_ALIAS items", async () => {
-    await request(URL)
+    await request(REST_URL)
       .get("/api/v2/aliases")
       .set("Authorization", `Bearer ${key}`)
       .trustLocalhost(true)
@@ -39,7 +58,7 @@ describe("aliases (v2)", () => {
   });
 
   it("alias entries expose name, query_url and metadata", async () => {
-    await request(URL)
+    await request(REST_URL)
       .get("/api/v2/aliases")
       .set("Authorization", `Bearer ${key}`)
       .trustLocalhost(true)
@@ -59,7 +78,7 @@ describe("aliases (v2)", () => {
   });
 
   it("does not include real (non-alias) queries", async () => {
-    await request(URL)
+    await request(REST_URL)
       .get("/api/v2/aliases")
       .set("Authorization", `Bearer ${key}`)
       .trustLocalhost(true)
@@ -73,7 +92,7 @@ describe("aliases (v2)", () => {
   });
 
   it("?all=false returns aliases from loaded modules only", async () => {
-    await request(URL)
+    await request(REST_URL)
       .get("/api/v2/aliases?all=false")
       .set("Authorization", `Bearer ${key}`)
       .trustLocalhost(true)
@@ -88,7 +107,7 @@ describe("aliases (v2)", () => {
     // able to read /aliases. This guards against accidentally widening
     // the default grant set.
     let legacyKey: string | undefined = undefined;
-    await request(URL)
+    await request(REST_URL)
       .get("/api/v2/login")
       .auth("legacy", "legacy-password")
       .trustLocalhost(true)
@@ -96,44 +115,44 @@ describe("aliases (v2)", () => {
       .then((response) => {
         legacyKey = response.body.key;
       });
-    await request(URL)
+    await request(REST_URL)
       .get("/api/v2/aliases")
       .set("Authorization", `Bearer ${legacyKey}`)
       .trustLocalhost(true)
       .expect(403);
   });
-});
 
-describe("aliases (v1)", () => {
-  let key: string | undefined = undefined;
+  describe("v1 alias of the same endpoint", () => {
+    let v1Key: string | undefined = undefined;
 
-  it("can login", async () => {
-    await request(URL)
-      .get("/api/v1/login")
-      .auth("admin", "default-password")
-      .trustLocalhost(true)
-      .expect(200)
-      .then((response) => {
-        key = response.body.key;
-      });
-  });
+    it("can login on /api/v1", async () => {
+      await request(REST_URL)
+        .get("/api/v1/login")
+        .auth("admin", "default-password")
+        .trustLocalhost(true)
+        .expect(200)
+        .then((response) => {
+          v1Key = response.body.key;
+        });
+    });
 
-  it("is mounted on /api/v1 as well", async () => {
-    await request(URL)
-      .get("/api/v1/aliases")
-      .set("Authorization", `Bearer ${key}`)
-      .trustLocalhost(true)
-      .expect(200)
-      .then((response) => {
-        expect(Array.isArray(response.body)).toBe(true);
-        const alias = response.body.find(
-          (a: { name: string }) => a.name === "mock_alias",
-        );
-        expect(alias).toBeDefined();
-        // v1 endpoint returns v1 query URLs.
-        expect(alias.query_url).toEqual(
-          "https://127.0.0.1:8443/api/v1/queries/mock_alias/",
-        );
-      });
+    it("is mounted on /api/v1 as well", async () => {
+      await request(REST_URL)
+        .get("/api/v1/aliases")
+        .set("Authorization", `Bearer ${v1Key}`)
+        .trustLocalhost(true)
+        .expect(200)
+        .then((response) => {
+          expect(Array.isArray(response.body)).toBe(true);
+          const alias = response.body.find(
+            (a: { name: string }) => a.name === "mock_alias",
+          );
+          expect(alias).toBeDefined();
+          // v1 endpoint returns v1 query URLs.
+          expect(alias.query_url).toEqual(
+            "https://127.0.0.1:8443/api/v1/queries/mock_alias/",
+          );
+        });
+    });
   });
 });
