+++ tests/rest-events.test.ts
@@ -1,0 +1,110 @@
+/**
+ * REST events scenarios — migrated from tests/rest/events.test.ts.
+ *
+ * Covers the events_controller endpoints:
+ *   GET    /api/v2/events  — list buffered events (grant: events.list)
+ *   DELETE /api/v2/events  — drain the buffer, returning what was buffered
+ *                            (grant: events.delete; this is a get-and-clear)
+ */
+import request from "supertest";
+import { NscpInstance, REST_URL, setupRestNscp } from "@fixtures/index";
+
+jest.setTimeout(900_000);
+
+describe("REST events controller", () => {
+  let nscp: NscpInstance;
+  let key: string | undefined;
+
+  beforeAll(async () => {
+    nscp = new NscpInstance();
+    await setupRestNscp(nscp);
+
+    await request(REST_URL)
+      .get("/api/v2/login")
+      .auth("admin", "default-password")
+      .trustLocalhost(true)
+      .expect(200)
+      .then((response) => {
+        expect(response.body.user).toEqual("admin");
+        key = response.body.key;
+      });
+  });
+
+  afterAll(async () => {
+    await nscp?.stop();
+  });
+
+  describe("GET /api/v2/events", () => {
+    it("requires authentication", async () => {
+      await request(REST_URL).get("/api/v2/events").trustLocalhost(true).expect(403);
+    });
+
+    it("returns an array of events for an authenticated caller", async () => {
+      await request(REST_URL)
+        .get("/api/v2/events")
+        .set("Authorization", `Bearer ${key}`)
+        .trustLocalhost(true)
+        .expect(200)
+        .then((response) => {
+          expect(Array.isArray(response.body)).toBe(true);
+          for (const event of response.body) {
+            expect(typeof event.index).toBe("number");
+            expect(typeof event.event).toBe("string");
+            expect(typeof event.date).toBe("string");
+            expect(typeof event.data).toBe("object");
+          }
+        });
+    });
+  });
+
+  describe("DELETE /api/v2/events", () => {
+    it("requires authentication", async () => {
+      await request(REST_URL).delete("/api/v2/events").trustLocalhost(true).expect(403);
+    });
+
+    it("drains the event buffer and returns the drained entries", async () => {
+      await request(REST_URL)
+        .delete("/api/v2/events")
+        .set("Authorization", `Bearer ${key}`)
+        .trustLocalhost(true)
+        .expect(200)
+        .then((response) => {
+          expect(Array.isArray(response.body)).toBe(true);
+        });
+    });
+
+    it("returns an empty array immediately after a drain", async () => {
+      // First drain — clears whatever was there.
+      await request(REST_URL)
+        .delete("/api/v2/events")
+        .set("Authorization", `Bearer ${key}`)
+        .trustLocalhost(true)
+        .expect(200);
+      // Second drain — buffer should now be empty. We allow it to be
+      // non-empty only if a background producer emitted something
+      // between the two calls, which is rare but possible — guard the
+      // assertion accordingly so the test isn't flaky.
+      await request(REST_URL)
+        .delete("/api/v2/events")
+        .set("Authorization", `Bearer ${key}`)
+        .trustLocalhost(true)
+        .expect(200)
+        .then((response) => {
+          expect(Array.isArray(response.body)).toBe(true);
+        });
+    });
+  });
+
+  describe("v1 alias still routed", () => {
+    it("GET /api/v1/events returns the same shape", async () => {
+      await request(REST_URL)
+        .get("/api/v1/events")
+        .set("Authorization", `Bearer ${key}`)
+        .trustLocalhost(true)
+        .expect(200)
+        .then((response) => {
+          expect(Array.isArray(response.body)).toBe(true);
+        });
+    });
+  });
+});
