--- tests/rest/.prettierrc
@@ -1,1 +1,0 @@
-{}
