--- tests/rest/nsclient.ini
@@ -1,45 +1,0 @@
-﻿[/modules]
-CheckSystem = disabled
-WEBServer = enabled
-CheckHelpers = enabled
-LUAScript = enabled
-
-[/settings/default]
-password = default-password
-allowed hosts = 127.0.0.1,::1
-
-[/settings/lua/scripts]
-mock = mock.lua
-
-[/settings/WEB/server/roles]
-view = *
-legacy = legacy,login.get
-full = *
-client = public,info.get,info.get.version,queries.list,queries.get,queries.execute,aliases.list,login.get,modules.list
-
-; Aliases registered via CheckHelpers - the alias endpoint test uses these to
-; verify QUERY_ALIAS items are returned and that QUERY items are not.
-[/settings/check helpers/alias]
-mock_alias = mock_query
-echo_alias = check_warning message=hello
-
-[/settings/WEB/server/users/client]
-role = client
-password = client-password
-
-[/settings/WEB/server/users/legacy]
-role = legacy
-password = legacy-password
-
-; The Jest suite authenticates as admin with the global default password
-; throughout (e.g. queries-v2.test.ts:12 -> auth("admin", "default-password")).
-; Pre-0.13 the WEBServer would silently overwrite the admin user's password
-; with /settings/default/password on every boot, masking whatever was here.
-; That overwrite was removed in 0.13 (commit 4fbe1aa1), so we now set the
-; admin password explicitly to match the value the tests use.
-[/settings/WEB/server/users/admin]
-role = full
-password = default-password
-
-[/settings/WEB/server/users/default-admin]
-role = full
