--- tests/rest/package.json
@@ -1,16 +1,0 @@
-{
-  "devDependencies": {
-    "@babel/core": "^7.28.6",
-    "@babel/preset-env": "^7.28.6",
-    "@babel/preset-typescript": "^7.28.5",
-    "@jest/globals": "^30.2.0",
-    "@types/supertest": "^6.0.3",
-    "babel-jest": "^30.2.0",
-    "jest": "^30.2.0",
-    "prettier": "3.8.0",
-    "supertest": "^7.2.2"
-  },
-  "scripts": {
-    "test": "jest"
-  }
-}
