+++ tests/src/global-setup.ts
@@ -2,12 +2,48 @@ import * as fs from "fs";
 import * as path from "path";
 import execa from "execa";
 
+/**
+ * Resolve a bare executable name (e.g. "nscp") against PATH. Returns the
+ * absolute path or undefined. Used so CI configs can set NSCP_BIN=nscp
+ * without having to know whether the install put the binary in /usr/sbin
+ * (Debian/RPM service binary convention) or /usr/bin.
+ */
+function whichOnPath(name: string): string | undefined {
+  const dirs = (process.env.PATH ?? "").split(path.delimiter);
+  const exts =
+    process.platform === "win32"
+      ? (process.env.PATHEXT ?? ".EXE;.CMD;.BAT").split(";")
+      : [""];
+  for (const dir of dirs) {
+    if (!dir) continue;
+    for (const ext of exts) {
+      const candidate = path.join(dir, name + ext);
+      if (fs.existsSync(candidate)) return candidate;
+    }
+  }
+  return undefined;
+}
+
 function findNscp(): string {
   if (process.env.NSCP_BIN) {
-    if (!fs.existsSync(process.env.NSCP_BIN)) {
-      throw new Error(`NSCP_BIN=${process.env.NSCP_BIN} does not exist`);
+    const raw = process.env.NSCP_BIN;
+    // Absolute / explicit-relative path — must exist as given.
+    if (path.isAbsolute(raw) || raw.includes(path.sep) || raw.includes("/")) {
+      if (!fs.existsSync(raw)) {
+        throw new Error(`NSCP_BIN=${raw} does not exist`);
+      }
+      return path.resolve(raw);
     }
-    return path.resolve(process.env.NSCP_BIN);
+    // Bare command name (e.g. NSCP_BIN=nscp) — resolve via PATH so the same
+    // env var works against a built tree, a Debian sbin install, and an RPM
+    // install without per-distro path hardcoding.
+    const resolved = whichOnPath(raw);
+    if (!resolved) {
+      throw new Error(
+        `NSCP_BIN=${raw} could not be resolved on PATH (PATH=${process.env.PATH ?? ""})`,
+      );
+    }
+    return resolved;
   }
   const buildDirs = [
     process.env.NSCP_BUILD_DIR,
@@ -40,7 +76,14 @@ function verifyDocker(): void {
 export default async function globalSetup(): Promise<void> {
   const bin = findNscp();
   process.env.NSCP_BIN = bin;
-  verifyDocker();
+  // The docker-using scenarios skip themselves via dockerOrSkip() when this
+  // env var is set, so probing for a daemon would just abort a perfectly
+  // valid no-docker run (e.g. the CI stage that only exercises the rest-*
+  // suites). Skip the probe and let the docker-needing scenarios — which
+  // won't run anyway — decide for themselves.
+  if (process.env.NSCP_SKIP_DOCKER !== "1") {
+    verifyDocker();
+  }
   // eslint-disable-next-line no-console
   console.log(`[integration] using nscp at ${bin}`);
 }
