+++ tests/src/index.ts
@@ -4,3 +4,4 @@ export * from "./tls";
 export * from "./files";
 export * from "./http";
 export * from "./log-on-fail";
+export * from "./rest-fixture";
